requirements.txt
python-telegram-bot==13.10
python-dotenv==0.19.2
requests==2.27.1
